don't erase (or try to) prefs
there's no point in resetting system settings. It won't help when switching from Electra and it only works with like 20% of the users for some reason. And when it does users complain about their settings (wifi, wallpaper etc) getting reset. An user even got a bootloop while doing this (judging from the fact that he/she saw a progress bar)